requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
openai>=1.20.0
pyyaml>=6.0
pytest>=8.0.0
httpx>=0.27.0
